feat(ion-button): add base styles for ionic theme buttons (#29653)
Issue number: internal

---------

<!-- Please do not submit updates to dependencies unless it fixes an
issue. -->

<!-- Please try to limit your pull request to one type (bugfix, feature,
etc). Submit multiple pull requests if needed. -->


## What is the new behavior?
<!-- Please describe the behavior or changes that are being added by
this PR. -->
This PR is being merged to ROU-10837 (ion-toolbar) where after a series
of PR's are merged, the final PR will be done to next branch. This was
done as most of these styles only make sense when used together in the
context of the ion-toolbar, but there were a lot of changes that made
sense to split in multiple PR's.

- Added new scss file for Ionic theme.
- Split the files between common and native.

## Does this introduce a breaking change?

- [ ] Yes
- [x] No

<!--
  If this introduces a breaking change:
1. Describe the impact and migration path for existing applications
below.
  2. Update the BREAKING.md file with the breaking change.
3. Add "BREAKING CHANGE: [...]" to the commit description when merging.
See
https://github.com/ionic-team/ionic-framework/blob/main/docs/CONTRIBUTING.md#footer
for more information.
-->